Here are the changes I made on the codebase.
1.Fixed the bug that I found wherein the numbers
gets saved on the array per clicked after inputitng the two pari of numbers.

I fixed it by moving the "sum = calculateSum()"
to the additionOperator event listener. this ensures that the inputted numbers
will only get saved on the array after clicking an operator button which later on be used to calculate the sum of the numbers.

2. Finished Task # 3 and Task on Numbers (1 to 9).