Here are the changes made on the codebase.
1. Enhance the logic of the numberButtons foreach() by adding a condition wherein the
the input display on the calculator resets when the previous previously displayed number
is not equal to the second / previously inputted number. this ensures
that when you input a third number, it will not include the
numbers of the second inputted number.

2. Refactor code on line 63 in order to display the accurate results on the calculator.
3. Added displayOfResult variable that gets the last sum of numbers
and the last number to be added on the equation.